rename files, prune api + README
