+++ lib/is-group-add-member.js
@@ -7,6 +7,9 @@ module.exports = Validator(schema, [
   oneGroupAtMost
 ])
 
+// TODO move to
+// spec/group/add-member/index.js
+
 function groupInFirstSlot ({ recps }) {
   return (
     isCloaked(recps[0]) ||
