+++ method/group/add-member.js
@@ -1,7 +1,8 @@
+const isValid = require('../../lib/is-group-add-member')
+
 module.exports = function GroupInvite (ssb, keystore) {
   return function groupInvite (groupId, authorIds, opts = {}, cb) {
     const { key, initialMsg } = keystore.group.get(groupId)
-    // const { groupKey, 
     // TODO cap authorIds to 15 (relevant to maxAttempts for unboxing)
 
     const content = {
@@ -10,16 +11,14 @@ module.exports = function GroupInvite (ssb, keystore) {
       groupKey: key.toString('base64'),
       initialMsg,
       tangles: {
-        // TODO figure out root + previous for whole group
         group: {
           root: initialMsg,
-          previous: [initialMsg]
+          previous: [initialMsg] // TODO calculate previous for whole group
         },
 
-        // TODO figure out previous for membership tangle (root is same as group.root)
         members: {
           root: initialMsg,
-          previous: [initialMsg]
+          previous: [initialMsg] // TODO calculate previous for members tangle
         }
       },
       recps: [groupId, ...authorIds]
@@ -27,7 +26,7 @@ module.exports = function GroupInvite (ssb, keystore) {
 
     if (opts.text) content.text = opts.text
 
-    // TODO validate before publish
+    if (!isValid(content)) return cb(new Error(isValid.errorsString))
 
     ssb.publish(content, cb)
   }
