+++ method/index.js
@@ -1,8 +1,8 @@
 module.exports = function Method (ssb, keystore, state) {
   return {
     group: {
-      create: require('./group/create')(ssb, keystore, state),
-      invite: require('./group/invite')(ssb, keystore, state)
+      init: require('./group/init')(ssb, keystore, state),
+      addMember: require('./group/add-member')(ssb, keystore, state)
     }
   }
 }
